+++ tests/testthat/test-extract_variables.R
@@ -52,8 +52,14 @@ test_that("extract_variables() ignores ifelse functions", {
   expect_equal(actual_vars, expected_vars)
 })
 
-test_that("extract_variables() ignores logical operators ", {
+test_that("extract_variables() ignores logical operators", {
   test_lhs      <- "test_var"
   equation      <- "ifelse(!(a>b)&a>c|c<b, a, b)"
   expect_equal(extract_variables(test_lhs, equation), c("a", "b", "c"))
+})
+
+test_that("extract_variables() ignores fixed delay", {
+  test_lhs <- "outflow"
+  equation <- 'sd_fixed_delay("inflow", 6, 2, 0, .memory)'
+  expect_equal(extract_variables(test_lhs, equation), "inflow")
 })