+++ tests/testthat/test-xml_model_components.R
@@ -82,8 +82,12 @@ test_that("create_level_obj_xmile() returns the expected object", {
     list(name = "initPopulation",
          value = "100"))
 
+  time_aux   <- list(name     = "time",
+                     equation = 0)
+
   level_obj    <- create_level_obj_xmile(test_stocks_xml,
-                                         test_vars, test_consts)
+                                         test_vars, test_consts,
+                                         time_aux = time_aux)
   actual_val   <- level_obj[[1]]
   expected_val <- list(name = "Population",
                        equation = "netGrowth",
@@ -107,8 +111,12 @@ test_that("create_level_obj_xmile() deals with levels with no flows", {
   test_vars   <- list()
   test_consts <- list()
 
+  time_aux   <- list(name     = "time",
+                     equation = 0)
+
   level_obj    <- create_level_obj_xmile(test_stocks_xml,
-                                         test_vars, test_consts)
+                                         test_vars, test_consts,
+                                         time_aux = time_aux)
   actual_obj   <- level_obj[[1]]
 
   expected_obj <- list(name = "Population",
@@ -143,8 +151,12 @@ test_that("create_level_obj_xmile() works should levels depend on other levels i
   test_vars   <- list()
   test_consts <- list()
 
+  time_aux   <- list(name     = "time",
+                     equation = 0)
+
   actual_obj    <- create_level_obj_xmile(test_stocks_xml,
-                                         test_vars, test_consts)
+                                         test_vars, test_consts,
+                                         time_aux = time_aux)
   expected_obj <- list(
     list(name      = "pop1",
          equation  = "net_change_1",
@@ -186,8 +198,12 @@ test_that("create_level_obj_xmile() returns the expected object in the presence 
 
   test_consts <- list()
 
+  time_aux   <- list(name     = "time",
+                     equation = 0)
+
   level_obj    <- create_level_obj_xmile(test_stocks_xml,
-                                         test_vars, test_consts)
+                                         test_vars, test_consts,
+                                         time_aux = time_aux)
   actual_val   <- level_obj[[1]]
   expected_val <- list(name = "Population",
                        equation = "births-emigration-deaths",
@@ -222,8 +238,12 @@ test_that("create_level_obj_xmile() returns the expected object in the presence 
 
   test_consts <- list()
 
+  time_aux   <- list(name     = "time",
+                     equation = 0)
+
   level_obj    <- create_level_obj_xmile(test_stocks_xml,
-                                         test_vars, test_consts)
+                                         test_vars, test_consts,
+                                         time_aux = time_aux)
   actual_val   <- level_obj[[1]]
   expected_val <- list(name = "Population",
                        equation = "births+immigration-deaths",
@@ -280,9 +300,14 @@ test_that("create_level_obj_xmile() takes into account builtin stocks", {
          initValue = 1)
     )
 
+  time_aux   <- list(name     = "time",
+                     equation = 0)
+
 
   actual_val      <- create_level_obj_xmile(test_stocks_xml, variables,
-                                            constants, builtin_stocks)
+                                            constants, builtin_stocks,
+                                            time_aux = time_aux)
+
   expected_val <- list(name      = "S1",
                        equation  = "adjust_S1",
                        initValue = 1)
@@ -327,10 +352,91 @@ test_that("create_level_obj_xmile() handles stock vectors", {
       equation  = "growth_B",
       initValue = 200))
 
-  actual_obj <- create_level_obj_xmile(test_stocks_xml, test_vars, test_consts)
+  time_aux   <- list(name     = "time",
+                     equation = 0)
+
+  actual_obj <- create_level_obj_xmile(test_stocks_xml, test_vars, test_consts,
+                                       time_aux = time_aux)
 
   expect_equal(actual_obj, expected_obj)
 })
+
+test_that("create_level_obj_xmile() handles DELAY FIXED from Vensim", {
+
+  test_xml <- xml2::read_xml('
+  <xmile version="1.0" xmlns="http://docs.oasis-open.org/xmile/ns/XMILE/v1.0">
+    <model>
+		  <variables>
+			  <stock name="out">
+				  <eqn>
+					out=
+	DELAY_FIXED(inflow, 2, 0)
+					</eqn>
+			  </stock>
+			  <stock name="test">
+			    <eqn>
+			    0
+					</eqn>
+					<inflow>
+					inflow
+					</inflow>
+					<outflow>
+					outflow
+					</outflow>
+			  </stock>
+      </variables>
+	  </model>
+  </xmile>')
+
+  test_stocks_xml <- xml2::xml_find_all(test_xml, ".//d1:stock")
+  test_vars   <- list()
+  test_consts <- list()
+
+  expected_obj <- list(
+    list(
+      name      = "test",
+      equation  = "inflow-outflow",
+      initValue = 0))
+
+  time_aux   <- list(name     = "time",
+                     equation = 0)
+
+  actual_obj <- create_level_obj_xmile(test_stocks_xml, test_vars, test_consts,
+                                       time_aux = time_aux)
+
+  expect_equal(actual_obj, expected_obj)
+})
+
+test_that("create_level_obj_xmile() handles step function in init value", {
+  stocks_xml <- list()
+
+  variables <- list(
+    list(name     = "orders",
+         equation = "3+ifelse(time>=2,3,0)"),
+    list(name     = "adjust_smoothed_step",
+         equation = "(orders-smoothed_step)/2")
+  )
+
+  constants <- list()
+
+  builtin_stocks <- list(list(name      = "smoothed_step",
+                              equation  = "adjust_smoothed_step",
+                              initValue = "orders"))
+
+  time_aux   <- list(name     = "time",
+                     equation = 0)
+
+  actual_obj <- create_level_obj_xmile(stocks_xml, variables, constants,
+                                       builtin_stocks, time_aux = time_aux)
+
+  expected_obj <- list(list(name      = "smoothed_step",
+                            equation  = "adjust_smoothed_step",
+                            initValue = 3))
+
+  expect_equal(actual_obj, expected_obj)
+})
+
+# extract_stock_info()----------------------------------------------------------
 
 test_that("extract_stock_info() handles a stock vector", {
   expected_obj <- list(